Added tests for firebase and signup
