+++ client/src/components/SignUp/SignUp.test.js
@@ -1,89 +1,96 @@
-import SignUp from '.'
 import React from 'react';
-import { render, fireEvent } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';
-import Enzyme, { shallow, mount } from 'enzyme';
+
+import { createMount } from '@material-ui/core/test-utils';
+import jest from 'jest-mock';
+import Enzyme from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 
-import { shape } from 'prop-types';
-
-// Instantiate router context
-const router = {
-  history: new BrowserRouter().history,
-  route: {
-    location: {},
-    match: {},
-  },
-};
-
-const createContext = () => ({
-  context: { router },
-  childContextTypes: { router: shape({}) },
-});
-
-// var firebasemock    = require('firebase-mock');
-
-// var mockauth = new firebasemock.MockAuthentication();
-// var mockdatabase = new firebasemock.MockFirebase();
-// var mockfirestore = new firebasemock.MockFirestore();
-// var mockstorage = new firebasemock.MockStorage();
-// var mockmessaging = new firebasemock.MockMessaging();
-// var mocksdk = new firebasemock.MockFirebaseSdk(
-//   // use null if your code does not use RTDB
-//   (path) => {
-//     return path ? mockdatabase.child(path) : mockdatabase;
-//   },
-//   // use null if your code does not use AUTHENTICATION
-//   () => {
-//     return mockauth;
-//   },
-//   // use null if your code does not use FIRESTORE
-//   () => {
-//     return null;
-//   },
-//   // use null if your code does not use STORAGE
-//   () => {
-//     return null;
-//   },
-//   // use null if your code does not use MESSAGING
-//   () => {
-//     return null;
-//   }
-// );
-
-// var jest = require('jest');
-
-// jest.mock('../path-to-firebase-init', () => {
-//     return mocksdk;
-// });
+import SignUp from '.'
+import {SignUpFormBase, styles} from './SignUpFormBase'
+import { withStyles } from '@material-ui/core/styles';
 
 Enzyme.configure({ adapter: new Adapter() })
 
-it('check if fields receive text', () => {
-    const { getByTestId } = render(<BrowserRouter><SignUp/></BrowserRouter>);
-    
-    const firstnameInput = getByTestId('firstnameInput');
-    const lastnameInput = getByTestId('lastnameInput');
-    const emailInput = getByTestId('emailInput');
-    const passwordOneInput = getByTestId('passwordOneInput');
-    const passwordTwoInput = getByTestId('passwordTwoInput');
-    
-    expect(firstnameInput).toHaveValue('');
-    expect(lastnameInput).toHaveValue('');
-    expect(emailInput).toHaveValue('');
-    expect(passwordOneInput).toHaveValue('');
-    expect(passwordTwoInput).toHaveValue('');
+describe('testing form', () => {
+  let mount;
   
-    fireEvent.change(firstnameInput, { target: { value: 'Mickey' } });
-    fireEvent.change(lastnameInput, { target: { value: 'Mouse' } });
-    fireEvent.change(emailInput, { target: { value: 'mickeymouse@disney.com' } });
-    fireEvent.change(passwordOneInput, { target: { value: '.waltdisneyrocks' } });
-    fireEvent.change(passwordTwoInput, { target: { value: '.waltdisneyrocks' } });
-
-    expect(firstnameInput).toHaveValue('Mickey');
-    expect(lastnameInput).toHaveValue('Mouse');
-    expect(emailInput).toHaveValue('mickeymouse@disney.com');
-    expect(passwordOneInput).toHaveValue('.waltdisneyrocks');
-    expect(passwordTwoInput).toHaveValue('.waltdisneyrocks');
+  beforeAll(() => {
+    mount = createMount({ });
   });
-  
+
+  it('check if fields receive text', () => {
+    const wrapper = mount(<BrowserRouter> <SignUp /> </BrowserRouter>);
+
+    expect(wrapper.find('input').at(0).value === '');
+    expect(wrapper.find('input').at(1).value === '');
+    expect(wrapper.find('input').at(2).value === '');
+    expect(wrapper.find('input').at(3).value === '');
+    expect(wrapper.find('input').at(4).value === '');
+
+    wrapper.find('input').at(0).simulate('change', { target: { name: 'firstname', value: 'Mickey' } });
+    wrapper.find('input').at(1).simulate('change', { target: { name: 'lastname', value: 'Mouse' } });
+    wrapper.find('input').at(2).simulate('change', { target: { name: 'email', value: 'mickeymouse@disney.com' } });
+    wrapper.find('input').at(3).simulate('change', { target: { name: 'passwordOne', value: '.waltdisneyrocks' } });
+    wrapper.find('input').at(4).simulate('change', { target: { name: 'passwordTwo', value: '.waltdisneyrocks' } });
+
+    expect(wrapper.find('input').at(0).value === 'Mickey');
+    expect(wrapper.find('input').at(1).value === 'Mouse');
+    expect(wrapper.find('input').at(2).value === 'mickeymouse@disney.com');
+    expect(wrapper.find('input').at(3).value === '.waltdisneyrocks');
+    expect(wrapper.find('input').at(4).value === '.waltdisneyrocks');
+  });
+});
+
+describe('testing button', () => {
+  let mount;
+
+  beforeAll(() => {
+    mount = createMount({ });
+  });
+
+  it('check if button disables correctly', () => {
+    const wrapper = mount(<BrowserRouter> <SignUp /> </BrowserRouter>);
+
+    // Check if every iteration still has the form disabled
+    expect(wrapper.find('button').at(0).prop('disabled')).toEqual(true);
+    wrapper.find('input').at(0).simulate('change', { target: { name: 'firstname', value: 'Mickey' } });
+    expect(wrapper.find('button').at(0).prop('disabled')).toEqual(true);
+    wrapper.find('input').at(1).simulate('change', { target: { name: 'lastname', value: 'Mouse' } });
+    expect(wrapper.find('button').at(0).prop('disabled')).toEqual(true);
+    wrapper.find('input').at(2).simulate('change', { target: { name: 'email', value: 'mickeymouse@disney.com' } });
+    expect(wrapper.find('button').at(0).prop('disabled')).toEqual(true);
+    wrapper.find('input').at(3).simulate('change', { target: { name: 'passwordOne', value: '.waltdisneyrocks' } });
+    expect(wrapper.find('button').at(0).prop('disabled')).toEqual(true);
+    wrapper.find('input').at(4).simulate('change', { target: { name: 'passwordTwo', value: '.waltdisneyrocks' } });
+    
+    // Check if valid input enables the button
+    expect(wrapper.find('button').at(0).prop('disabled')).toEqual(false);
+
+    // Check if mismatched passwords disables the button
+    wrapper.find('input').at(4).simulate('change', { target: { name: 'passwordTwo', value: '.waltdisneyrocls' } });
+    expect(wrapper.find('button').at(0).prop('disabled')).toEqual(true);
+  });
+
+  it('testing button click', () => {
+    const SignUpFormBaseStyled = withStyles(styles)(SignUpFormBase);
+    const wrapper = mount(<BrowserRouter> <SignUpFormBaseStyled/> </BrowserRouter>);
+    const buttonClick = jest.fn().mockImplementation((event) => { return "clicked" });
+    wrapper.instance().onSubmit = buttonClick;
+    wrapper.update();
+    
+    wrapper.find('input').at(0).simulate('change', { target: { name: 'firstname', value: 'Mickey' } });
+    wrapper.find('input').at(1).simulate('change', { target: { name: 'lastname', value: 'Mouse' } });
+    wrapper.find('input').at(2).simulate('change', { target: { name: 'email', value: 'mickeymouse@disney.com' } });
+    wrapper.find('input').at(3).simulate('change', { target: { name: 'passwordOne', value: '.waltdisneyrocks' } });
+    wrapper.find('input').at(4).simulate('change', { target: { name: 'passwordTwo', value: '.waltdisneyrocks' } });
+    
+    const buttonInstance = wrapper.find('button').at(0);
+    
+    buttonInstance.props().onClick = buttonClick;
+    const eventMock = { preventDefault: jest.fn() };
+    buttonInstance.props().onClick(eventMock);
+    expect(buttonClick).toBeCalledWith(eventMock);
+  });
+});
+
