+++ client/src/components/SignUp/index.js
@@ -1,246 +1,15 @@
-import React, { Component } from 'react';
+import React from 'react';
 import { Link, withRouter } from 'react-router-dom';
 import { compose } from 'recompose';
 
-import Button from '@material-ui/core/Button';
-import CssBaseline from '@material-ui/core/CssBaseline';
-import TextField from '@material-ui/core/TextField';
-import FormControlLabel from '@material-ui/core/FormControlLabel';
-import Grid from '@material-ui/core/Grid';
-import Typography from '@material-ui/core/Typography';
 import { withStyles } from '@material-ui/core/styles';
-import Container from '@material-ui/core/Container';
-import Card from '@material-ui/core/Card'
-import CardContent from '@material-ui/core/CardContent'
-
-import { SignInLink } from '../SignIn';
 import { withFirebase } from '../Firebase';
+import {SignUpFormBase, styles} from './SignUpFormBase'
 import * as ROUTES from '../../constants/routes';
-
-const styles = (theme) => ({
-  card: {
-    marginTop: 50,
-    maxWidth: 500,
-    margin: "auto",
-    transition: "0.3s",
-    boxShadow: "0 8px 40px -12px rgba(0,0,0,0.3)",
-    "&:hover": {
-      boxShadow: "0 16px 70px -12.125px rgba(0,0,0,0.3)"
-    }
-  },
-  paper: {
-    marginTop: theme.spacing(8),
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-  },
-  form: {
-    width: '100%', // Fix IE 11 issue.
-    marginTop: theme.spacing(3),
-  },
-  submit: {
-    margin: theme.spacing(3, 0, 2),
-  },
-  error: {
-    color: 'red',
-    fontSize: 'small',
-    margin: theme.spacing(3, 0, 2),
-  }
-});
 
 const SignUpPage = () => (
   <SignUpForm />
 );
-
-const INITIAL_STATE = {
-  firstname: '',
-  lastname: '',
-  email: '',
-  passwordOne: '',
-  passwordTwo: '',
-  error: null,
-};
-
-class SignUpFormBase extends Component {
-  constructor(props) {
-    super(props);
-    this.state = { ...INITIAL_STATE };
-  }
-  onChange = event => {
-    this.setState({ [event.target.name]: event.target.value });
-  };
-  
-  onSubmit = event => {
-    console.log("OnSubmit Called")
-    event.preventDefault();
-    const { firstname, lastname, email, passwordOne } = this.state;
-    this.props.firebase
-      .doCreateUserWithEmailAndPassword(email, passwordOne)
-      .then(authUser => {
-        // Get user information
-        return this.props.firebase
-          .user(authUser.user.uid)
-          .set({
-            firstname,
-            lastname,
-            email,
-          });
-      })
-      .then(authUser => {
-        this.setState({ ...INITIAL_STATE });
-        this.props.history.push(ROUTES.LANDING);
-      })
-      .catch(error => {
-        console.log('Error: ', error.message);
-        this.setState({ error });
-      });
-  };
-
-  render() {
-    const { classes } = this.props;
-
-    const {
-      firstname,
-      lastname,
-      email,
-      passwordOne: passwordOne,
-      passwordTwo: passwordTwo,
-      error,
-    } = this.state;
-
-    const isInvalid =
-      passwordOne !== passwordTwo ||
-      passwordOne === '' ||
-      email === '' ||
-      firstname === '' ||
-      lastname === '';
-
-    return (
-      <div id="centered-masthead">
-      <Card className={classes.card}>
-        <CardContent>
-          <Container component="main" maxWidth="xs">
-            <CssBaseline />
-            <div className={classes.paper}>
-              <Typography component="h1" variant="h5">
-                Sign up
-              </Typography>
-              <form className={classes.form} noValidate>
-                <Grid container spacing={2}>
-                  <Grid item xs={12} sm={6}>
-                    <TextField
-                      inputProps={{
-                        'data-testid': 'firstnameInput'
-                      }}
-                      autoComplete="fname"
-                      name="firstname"
-                      value={firstname}
-                      onChange={this.onChange}
-                      variant="outlined"
-                      required
-                      fullWidth
-                      id="firstname"
-                      label="First Name"
-                      autoFocus
-                    />
-                  </Grid>
-                  <Grid item xs={12} sm={6}>
-                    <TextField
-                      inputProps={{
-                        'data-testid': 'lastnameInput'
-                      }}
-                      variant="outlined"
-                      required
-                      fullWidth
-                      id="lastname"
-                      label="Last Name"
-                      name="lastname"
-                      value={lastname}
-                      onChange={this.onChange}
-                      autoComplete="lname"
-                    />
-                  </Grid>
-                  <Grid item xs={12}>
-                    <TextField
-                      inputProps={{
-                        'data-testid': 'emailInput'
-                      }}
-                      variant="outlined"
-                      required
-                      fullWidth
-                      id="email"
-                      label="Email Address"
-                      name="email"
-                      value={email}
-                      onChange={this.onChange}
-                      name="email"
-                      autoComplete="email"
-                    />
-                  </Grid>
-                  <Grid item xs={12}>
-                    <TextField
-                      inputProps={{
-                        'data-testid': 'passwordOneInput'
-                      }}
-                      variant="outlined"
-                      required
-                      fullWidth
-                      name="passwordOne"
-                      label="Password"
-                      type="password"
-                      id="password"
-                      value={passwordOne}
-                      onChange={this.onChange}
-                      autoComplete="current-password"
-                    />
-                  </Grid>
-                  <Grid item xs={12}>
-                    <TextField
-                      inputProps={{
-                        'data-testid': 'passwordTwoInput'
-                      }}
-                      variant="outlined"
-                      required
-                      fullWidth
-                      name="passwordTwo"
-                      label="Confirm Password"
-                      type="password"
-                      id="password"
-                      value={passwordTwo}
-                      onChange={this.onChange}
-                      autoComplete="current-password"
-                    />
-                  </Grid>
-                </Grid>
-                <Button
-                  disabled={isInvalid}
-                  onClick={this.onSubmit}
-                  variant='primary'
-                  type="submit"
-                  fullWidth
-                  variant="contained"
-                  color="primary"
-                  className={classes.submit}
-                >
-                  Sign Up
-                </Button>
-                
-                {error && <Typography className={classes.error}>{error.message}</Typography>}
-
-                <Grid container justify="flex-end">
-                  <Grid item>
-                    <SignInLink/>
-                  </Grid>
-                </Grid>
-              </form>
-            </div>
-          </Container>
-        </CardContent>
-      </Card>
-      </div>
-    );
-  }
-}
 
 const SignUpLink = () => (
   <p>
